Refactor interfaces for useUser and useUi hooks
