+++ src/types/index.ts
@@ -2,3 +2,8 @@ export interface UserCredentials {
   email: string;
   password: string;
 }
+
+export interface ErrorData {
+  message: string;
+  isError: boolean;
+}
